+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.3] - 2023-04-13
+
+### Changed
+
+- Upgrade CDK to version 2
+- Upgrade nodejs version for build pipeline to 16
+- Upgrade nodejs version for lambda functions to 16
+- Include ClusterSize in deployment metrics
+- Upgrade aws-sdk version to mitigate [CVE-2023-0842: xml2js vulnerable to prototype pollution](https://nvd.nist.gov/vuln/detail/CVE-2023-0842)
+
 ## [4.0.2] - 2022-09-14
 
 ### Changed